+++ document.lib.rest/GlobalUsings.cs
@@ -7,7 +7,6 @@ global using document.lib.rest.Api.Contract;
 global using document.lib.rest.Api.Definition;
 global using document.lib.shared.Extensions;
 global using document.lib.shared.Interfaces;
-global using document.lib.shared.Models.Models;
 global using Microsoft.AspNetCore.Authentication.JwtBearer;
 global using Microsoft.AspNetCore.Mvc;
 global using Microsoft.EntityFrameworkCore;
