+++ document.lib.shared/Interfaces/ICategoryRepository.cs
@@ -1,6 +1,4 @@
-﻿using document.lib.shared.Models.Models;
-
-namespace document.lib.shared.Interfaces;
+﻿namespace document.lib.shared.Interfaces;
 
 public interface ICategoryRepository<T>
 {
