+++ document.lib.shared/Interfaces/ICategoryService.cs
@@ -1,5 +1,5 @@
 ﻿using document.lib.shared.Enums;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 
 namespace document.lib.shared.Interfaces;
