+++ document.lib.shared/Interfaces/IRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace document.lib.shared.Interfaces;
+
+public interface IRepository
+{
+    Task SaveAsync();
+}
