+++ document.lib.shared/Interfaces/Models/IFolderModelExtensions.cs
@@ -1,0 +1,9 @@
+﻿namespace document.lib.shared.Interfaces.Models;
+
+// Interface for FolderModel
+
+public interface IFolderModelExtensions
+{
+    bool IsActive { get; set; }
+    int? GetLastRegisterNumber();
+}
