+++ document.lib.shared/Models/Update/DocumentUpdateModel.cs
@@ -1,5 +1,5 @@
 ﻿using document.lib.ef.Entities;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 
 namespace document.lib.shared.Models.Update;
