+++ document.lib.shared/Repositories/Cosmos/DocumentCosmosRepository.cs
@@ -1,7 +1,7 @@
 ﻿using document.lib.shared.Constants;
 using document.lib.shared.Interfaces;
 using document.lib.shared.Models;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 using document.lib.shared.TableEntities;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Options;
