+++ document.lib.shared/Repositories/Sql/DocumentSqlRepository.cs
@@ -1,7 +1,7 @@
 ﻿using document.lib.ef;
 using document.lib.ef.Entities;
 using document.lib.shared.Interfaces;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 using Microsoft.EntityFrameworkCore;
 
 namespace document.lib.shared.Repositories.Sql;
