+++ document.lib.shared/Services/DocumentSqlService.cs
@@ -2,7 +2,7 @@
 using Azure.Storage.Blobs.Models;
 using document.lib.ef.Entities;
 using document.lib.shared.Interfaces;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 using document.lib.shared.Models.Update;
 
 namespace document.lib.shared.Services
