+++ document.lib.shared/Services/FolderSqlService.cs
@@ -1,6 +1,6 @@
 ﻿using document.lib.ef.Entities;
 using document.lib.shared.Interfaces;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 using document.lib.shared.Models.Update;
 
 namespace document.lib.shared.Services;
@@ -27,7 +27,6 @@ public class FolderSqlService(IFolderRepository<EfFolder> folderRepository)
     {
         var folder = await folderRepository.GetFolderAsync(id);
         if (folder == null) return null;
-
         return Map(folder);
     }
