+++ document.lib.shared/Services/TagSqlService.cs
@@ -1,6 +1,6 @@
 ﻿using document.lib.ef.Entities;
 using document.lib.shared.Interfaces;
-using document.lib.shared.Models.Models;
+using document.lib.shared.Models.Data;
 
 namespace document.lib.shared.Services;
