+++ document.lib.web/Pages/DocumentOverview.razor
@@ -2,7 +2,7 @@
 @using document.lib.shared.Services
 @using System.Web
 @using document.lib.shared.Interfaces
-@using document.lib.shared.Models.Models
+@using document.lib.shared.Models.Data
 
 @inject CosmosQueryService _cosmosQueryService
 @inject NavigationManager NavigationManager
