+++ document.lib.web/Pages/EditDocument.razor
@@ -5,7 +5,7 @@
 @using document.lib.shared.TableEntities
 @using document.lib.shared.Models
 @using document.lib.shared.Interfaces
-@using document.lib.shared.Models.Models
+@using document.lib.shared.Models.Data
 
 @inject CosmosQueryService _cosmosQueryService
 @inject CosmosMetadataService _cosmosMetadataService
