+++ document.lib.web/Pages/FolderOverview.razor
@@ -1,6 +1,6 @@
 ﻿@page "/folderoverview"
 @using document.lib.shared.Interfaces
-@using document.lib.shared.Models.Models
+@using document.lib.shared.Models.Data
 
 @inject IFolderService _folderService;
