+++ document.lib.web/Pages/NewDocuments.razor
@@ -2,7 +2,7 @@
 @using document.lib.shared.Services
 @using System.Web
 @using document.lib.shared.Interfaces
-@using document.lib.shared.Models.Models
+@using document.lib.shared.Models.Data
 
 <h3>New Documents</h3>
